Move modal components, rename and fix imports
