+++ src/components/modals/index.js
@@ -1,0 +1,2 @@
+export { default as ModalShell } from '~/components/modals/ModalShell';
+export { default as ConfirmModalBody } from '~/components/modals/ConfirmModalBody';
